Fixes for TRAFODION-1898 and TRAFODION-1894.
The Client Install guide has been updated to support Windows 10. In
addition, the content has been tested and changed accordingly. Finally, I
added a blurb about how to handle the fact that the WinODBC driver is not
part of the download package. (TRAFODION-1898)

The Command Interface guide was updated to used the correct default port
number (23400) and to allow code examples to be downloaded.

The documentation page was updated to link to select wiki pages for UDF
and manageability.

The download page was updated to clarify the content of the Client
download package. (TRAFODION-1894).

Please rebuild the 1.3.0 and 2.0.0 documentation. Once done, please
refresh the Trafodion web site.