[TRAFODION-1918] Remove auto flush concept in Trafodion
The auto flush is on by default in insertRows method. But insertRows was
incorrectly called with the wrong parameter. In fixing this issue, it was
observed that auto flush is deprecated in HBase 1.x. It is also incorrect
to disable auto flush in Trafodion. Hence removed the code related to
auto flush, flush table and flush all tables in Trafodion.

All these are dormant code in Trafodion.

Also disabled executor/TEST106 for now because it seems to cause a hang
at the time of drop table.