Fixed a problem that prevented the built document from being published.
The causes of this issue is:

1. The <modules> entry is overkill when making changes to the website
only.
2. The DCS and REST documents do not build on Windows.

So, I normally comment out the <modules> entry in the main pom.xml for the
reasons above. Not optimal since it's possible to forget to uncomment.
This time, I did and no document was published (the post-site step).

Please rerun the steps needed to build and publish the 1.3.0 and 2.0.0
documentation.