added enhanced check to handle default values copy
Current code has an optimization to move source default
value to target if total length (null+VClen+length) match.
This is sufficient for current scenarios where the difference
between source and target is only in VClen indicator (2 or 4 bytes).
But that assumption may not be true in future and was pointed out
by reviewer. An enhanced check has been added to make sure
that all sub-parts (nullLen, vcLen, dataLen) also match.