[TRAFODION-1929] Set correct working dir for phx dependency builds
Per the Jira, the testing environment now invokes the test script in
the trafodion user environment, but that also has the side effect of
making the current directory as trafodion home directory.

The test script already calculates the path where it was invoked,
so changing directory there makes the dependent builds work.