TRAFODION-1940 better NATable statistics plus OID, cleanhb fixes
TRAFODION-1939 Remove OID from the list of reserved words
TRAFODION-1940 Improve NATable statistics
                - Add the context of the NATable cache
                  to the natablecache() UDF
                - Add the entry size to the
                  natablecacheentries() UDF

Also changed the ilh_cleanhb script to print a help text
on how to use it and require a parameter 0, 1 or 2.
Still leaving the old, misspelled directory for zookeeper
in the script until everyone has moved to a later version
that puts the zookeeper data under local_hadoop/data/zookeeper.