Misc cleanup items
Added a debugging method to print information for the list of all
Trafodion IPC connections in a process - that method can be called
from the debugger.

Removed unneeded code in preCodeGen - ValueIdSets for predicates imply
an AND, so no AND operators must be found in them.

Removed unnecessary flags in a makefile.