[TRAFODION-1956] Timeout for arkcmp process
A new CQD COMPILER_IDLE_TIMEOUT is added to control the compiler
timeout duration. Default value is 1800 seconds.

The compiler processes are killed when a SQL session ends at the
time of ODBC/JDBC disconnection. For an active session the compiler
process is killed when a compiler process remains idle more
than the compiler idle timeout.

The process in session needs to do some SQL for this feature to
kick in. For the idle connected process, the connection idle
timeout will enable killing compiler process. The compiler idle
timeout will be honored but it would take at least connection
idle timeout duration.