TRAFODION-2002 check columns for Hive inserts
Also TRAFODION-1904 column list in insert/select into Hive table is ignored

Added checks for matching number and data types, so that the
data to be inserted matches the column layout of the Hive target
table.

Using a column list for insert into a Hive table now will raise
SQL error 4223 instead of being ignored (so it is still not supported).