[TRAFODION-2021] Internal error on upsert with index
In this particular case, the extra nodes added by index maintenance
and the WHERE clause interacted in a way that caused some missing
characteristic inputs, leading to the generator error.  The fix is to
add the required inputs not directly to the nodes, but as "outer
references" to the BindScope, so that they get added automatically to
any nodes created in that BindScope. Since the predicates get created
by a method called at a higher level, I also needed to add a "remote
control" feature for the BindScope of a MergeUpdate node.

Also: [TRAFODION-2026] Upsert fails on table with nullable key column.

When creating the predicate that links the target clustering key with
the corresponding source values, use "special nulls" semantics for
nullable key columns, so a comparesion NULL = NULL is allowed and
yields TRUE.