TRAFODION [109] Instrument Trafodion to work with Secure Hadoop (Kerberos)
The first phase of automation with Kerberos is ready.

The following use cases are supported

I.	New secure install: Customer installs Trafodion on a cluster with
 secure Hadoop. There is no existing Trafodion installation.

II.	Conversion from non-secure to secure: Customer has Trafodion installed
on a non-secure Hadoop cluster. Customer enables Hadoop security on this
existing cluster. Customer runs a Trafodion script to enable Hadoop security
integration with Trafodion.

See the JIRA for details on the changes.