[TRAFODION-2126] Add OS version and processor type to package names
Add RH6-x86_64 to the server and client package names. This gives us
differentiation when we start building on RedHat 7 (TRAFODION-1861).

I also fixed the setting of JAVA_HOME in CentOS 7 environment, which
does not include x86_64 in the path name.

The build does not yet work on CentOS/RedHat 7, due to compiler
differences (TRAFODION-1868).