[TRAFODION-2129] Trafodion to avoid use of deprecated HBase APIs/Classes
References to the deprecated classes HConnection, HConnectionManager and
HBaseAdmin are reworked to use Connection, ConnectionFactory and Admin classes
respectively.

A connection object is created and it is passed around to ensure that the
same connection objects within the same process.

Admin object is created and closed after its use immediately.