[TRAFODION-2138] Hive scan on wide tables can result in lost rows or error
The fix has three parts
a)	RangeTailIOSize will now default to maximum length of a row, if > 16KB.
Each esp range will look ahead by this size. Previously this was hard coded to 16KB.
b)	If a whole buffer does not have the start of a row, that is not an error.
c)	We raise an error during compile if maximum row size is greater than size of IO buffer.
We need to have the whole row (from start to finish) in a contiguous buffer to do the rest of our logic.
Currently IO buffer defaults to 64 MB.