Various fixes, details listed below
--error message returned during string to float conversion error was not
correct. That has been fixed.

-- ordering was not getting set correctly during key encoding.

-- stored descriptor was not getting stored correctly if
the generated length was exactly 10000 bytes (max length of TEXT.text col).

-- time part of european date format is now handled correctly
by using '.' as delimiter instead of ':'

-- repeat function was crashing if specified maxlength was not sufficient
to hold the result. It now returns an overflow error.

-- tinyint key columns were not handling the case of overflow/underflow
when used in a key predicate. That has been fixed.

-- Numeric datatype with precision less than 3 is now created as tinyint
and handled in expressions.

-- some non-tinyint datatypes(float, decimal) were not being handled
correctly when converted to tinyint. That has been fixed.

-- some more pcode instructions involving tinyint range check and
conversions have been added.