Take 5 for javadoc change - taking a completely different approach
After talking to Steve and Gunnar, I am implementing a different
method for adding apidocs. We can leave the generation out of the
docs build and have a separate step to generate apidocs. The
new file introduced here, build_apidocs.sh will be used for the
separate step. Removing the code that invokes build_apidocs.sh
from the docs build.