TRAFODION-2203 - a user can grant privileges that he doesn’t have ...
... to other users/roles successfully

In this case, the user/role did not get the privilege requested even though the
operation successfully completed.  So the requester is lead to believe that the
privilege was granted.

ANSI states that: "warning <privilege not granted>" should be displayed for
each combination of grantee<=>privilege that was not granted. However,
privileges that can be successfully granted should be granted. The grant code
does not grant any privileges it cannot grant but is not reporting warnings if
the privilege is not granted. Ditto for revoke.

The code now reports warnings if not all privileges were granted or revoked for
both object and column privileges.

Also As part of this fix, the next piece of unifying object and column
privileges has been performed.  This task:

- Replaced ColPrivEntry with a PrivMgrCoreDesc - now object and column privs
  have the same base structure.
- Create a new method that performs common functions between grant and revoke
  statements
- Removed methods not longer needed
- Use column level privileges in the privsToGrant and privsToRevoke structs
- Fixed bug in showddl where privileges were not always displayed.
- Minor changes to make object and columns names more unified