[TRAFODION-2159] Promoting outputs of child with MapValueId node
Promoting outputs of child was not working correctly when the child is a
MapValueId node. Since the promoted outputs are not part of the map, MVI
could not produce the promoted outputs though its child was willing to.
Also fixed another bug recently introduced in the moveUpGrbyTransformation,
when there is a subquery groupby in the correlated subquery.