[TRAFODION-2262] Mxosrvr or java core with the stack trace pointing to log4Cxx functions
The Logger repository needs to be initialized. The method QRLogger::initLog4cxx
initialized the commonLogger infrastructure also.

PR #745 merged to fix this issue exposed a problem in SSMP. This
caused SSMP to dump core when a node is brought down.
SSMP should have create the connection to SSCP on that node.