[TRAFODION-2314] MXOSRVR sometimes exit abnormally with NAMutex assert
The test suite fails becuase pthread_mutex_unlock returns EPERM error

pthread_mutex_lock and pthread_mutex_unlock always return success when
a program is not linked with -lpthread option. It is linked this way
when a multi-threaded program is run in single thread mode to avoid the
overhead of locking.

The mxosrvr program is now linked with -lpthread. However, it is not
clear how and why it would fix the EPERM error returned from
pthread_mutex_unlock API. But it is good to ensure that mxosrvr .sos and
the binaries are compiled with -lpthread.

Will revisit this issue if the problem persists.