[TRAFODION-2291] Ambari integration
Integrates trafodion with Ambari cluster manager via an mpack (management
pack) plug-in.

As described in install/README.md file, to leverage this feature, two RPMs
are built. One to install the mpack and one used by Amari to install trafodion
itself. The mpack plug-in is built with a URL that points to where the
trafodion RPM is hosted (a yum repo). It can be specified on the make command
line as REPO_URL=...