TRAFODION-2439 add support for traf views on hive tables
support for creation of trafodion views on hive tables.
cqd hive_views is set to ON by default to enable it.

During view creation, an implicit external table is created
on underlying hive table if one doesn't already exists.
This is used to track usage info
in views_usage metadata table and to track privileges.
Usage info is used to return views on tables or tables in views
during GET stmts.

trafodion view is created in trafodion catalog/schema and not in hive
catalog/schema.

delete and update operations are not allowed on traf/hive views.

drop of external hive table with CASCADE option will cause all
dependent views to be dropped.
drop of external table without CASCADE option will not drop it
an a dependent view exists on that table.

If the underlying hive table is dropped from hive shell,
then an error is returned at runtime when that view is accessed.

new test regress/hive/TEST007 contains tests for this feature.