[TRAFODION-2452] Fix upgrade or removal of Ambari mpack
Enhance traf_ambari RPM to call upgrade-mpack instead of install-mpack
when we are upgrading the mpack.  Unfortunately, the upgrade-mpack command
does not (yet) seem to do the needed things.  More tweaks may be needed
in the future to support upgrade from trafodion 2.1.0 to later release.

Also support removal of the RPM to clean up everything.