[TRAFODION-2351] Bulk load with log error rows enhancements
Load with log error rows won't work if there are multiple ranges
involved in the hdfs scan. The hdfs log file is closed when the
first range of hdfs scan is done.  The error rows in the subsequent
ranges will not be logged. Instead an SQL warning is generated for
each process.  When these warnings are shipped from ESPs to master,
ESP dumps core.

Added tests to test load with log error rows and continue on error