TRAFODION-2175 a user should only see specific schemas/tables that he has privs TRAFODION-1573 Additional GET commands for privileges
This is one of several deliveries to improve GET commands:
- return details about objects that are only visible to the current_user.
- support additional statements

This delivery adds visibility checking to to the following commands:

  get users;
  get roles;
  get users for role <role>;
  get roles for user <user>;

It now supports the following GET statements:

  get privileges for user <user>;
  get privileges for role <role>;

When getting privileges for a user or role, the output is as follows:

  <privileges>    <object> <optional column name>

<privileges> output is a character string containing granted privileges:

  char 1: SELECT_PRIV     - "S"
  char 2: INSERT_PRIV     - "I"
  char 3: DELETE_PRIV     - "D"
  char 4: UPDATE_PRIV     - "U"
  char 5: USAGE_PRIV      - "G"
  char 6: REFERENCES_PRIV - "R"
  char 7: EXECUTE_PRIV    - "E"

If the user/role has no privilege then "-" is specified for the character

Example:  get privileges for user userx;

   SI--G-E    TRAFODION.MYSCH.MYTABLE1
   S------    TRAFODION.MYSCH.MYTABLE2 <Column> column1
   S--U---    TRAFODION.MYSCH.MYTABLE2 <Column> column2

When getting privileges for a role, only those privileges granted to the role
are returned.

When getting privileges for a user, privileges directly granted to the user
plus privileges granted to any roles associated with the user are returned.

A new regression test - privs1/TEST123 has been added to test these changes.

In addition, unused methods were removed from Get code.