[TRAFODION-2456] Improve documentation for UDF parallelism
Just setting the desired degree of parallelism often doesn't
actually choose a parallel plan, the UDF writer may also need
to specify an estimate of the number of rows produced.

Also made the methods that are not intended for the UDR writer
package-private (no qualifier). This has two good side-effects: One is
that UDR writers no longer can call these methods (accidentally) and
the other is that we can set Javadoc to display only public methods,
which suppresses these package-private methods.