[TRAFODION-2457] Re-enable compGeneral/TEST042 and try to stabilize
Implemented suggestions made by Suresh and Selva. Suresh found out
that the test gives different results when authorization is
initialized, and that caused problems because it is usually
initialized on developer workstations but not on Jenkins.

Suresh and Selva suggested to suppress the key length, since that can
vary with different CQD settings like SCHEMA.

Added the queries to the log and expected files to see whether that
will make failure diagnosis easier. The queries get filtered out, so
should not cause test failures.