TRAFODION-2175 a user should only see specific schemas/tables that he has privs
Fixed a problem with regression test Hive/TEST009.

The failed test is showing an additional table being stored in NATable cache.
If authorization is enabled, then "get tables" retrieves the list of roles
assigned the current user.  This causes an additional table to be cached.
However, if authorization is off, the role check is not made so the number of
entries in cache is less.  This causes a regressions failure depending on the
environment.

Since this is a test for caching, removed the get table commands from the test to
avoid the mismatch.