TRAFODION-2534 CLASSPATH for Java UDRs
Two changes in the custom Java class loader used for UDRs.

First, the class loader now adds any jar files stored in
$TRAF_HOME/udr/public/external_libs to its CLASSPATH.
The choice of this directory is anticipating work on
isolated UDFs, see TRAFODION-2561.

Second, the custom class loader LmClassLoader now uses
java.net.URLClassLoader to do most of the work. This allows
us to delete a lot of code in this file. When this class
was first written, many years ago, the URLClassLoader was
probably not yet available.

Finally, with the new class loader, the JDBC predefined
UDF can also be simplified.