[TRAFODION-2569] Improve handling of index hints
Allowing hints in update and delete statements - they apply to the
scan of the table. Also making sure that heuristics to eliminate
indexes don't apply to indexes listed in a hint. Excluding the right
child of an index join when giving priority to plans that satisfy the
hint. Otherwise, an index join (index - base table) could be chosen
when the hint lists the base table.

Also removed some unused interfaces:
- no_log option in grammar, since it made changing the update syntax harder.
  If we decide to implement incremental MVs in the future, we should use
  a better parser or a different syntax for the no log feature.
- Some obsolete parser rules.
- Another set of hints in the scan node, in addition to those in the RelExpr.