[TRAFODION-1906] Dop is reported incorrectly in RMS operator stats
DOP should be set to 1 in the ExOperStats constructors when TDB is passed
in. DOP is reset to 0, when the statistics area in the ESP processes
are shipped to higher tier process to avoid double counting.
DOP is restored back to 1 before the next execution again.

scratchIOSize is now initialized to 0 in the BMOStats constructor
to avoid junk value seen at times.