[TRAFODION-2611] TMUDF with input parameters returns wrong answer
When a TMUDF had a query with an input parameter or a CURRENT_TIMESTAMP
(or CURRENT_DATE, CURRENT_TIME) in it, we sometimes return incorrect
values.

The problem was that these values were not set up correctly in the
dataflow and they were not shipped as characteristic inputs to the
operators that were referring to them.