[TRAFODION-2615] TMUDF returns wrong results with small numeric inputs
The problem is that the new TINYINT type sometimes is stored in 1
byte, other times in 2 bytes. The TMUDF code makes the assumption that
we can compute the storage size from the precision. The fix is
in the binder and makes sure the storage size matches the storage size
that the TMUDF code will compute.