[TRAFODION-2616] Nested join regression after fix for TRAFODION-2569
When pushing predicates into a scan node, it is not ok to copy the
index info of the original scan to the new scan node, since the
additional predicates may change the set of qualifying indexes.

Also, the method to copy the index info needs to copy the set of
computed predicates, since these are computed in the addIndexInfo()
method.