Fix for Memory leak in compiler
Small category  (less than 24 characters) is accommodated as union of data members to store
the metadata of large or medium category. Hence small category need not be deallocated.
Medium category (strings between 24 and 254 characters) and Large category (above 254) are allocated
from heap. In the original code, the medium category should have been deallocated as soon as NAString is
deallocated.  But this code is not effective because medium category is commented out. The large category
keeps track of reference count and it is deallocated when the reference count becomes zero.

Without the change in FBString.h the medium category is considered as a small category. With my change,
the medium category strings are now considered as large category and will undergo reference count concepts.
Hence it will be deallocated assuming the reference counts work well.

In addition, the most commonly used function CharType::generateTextThenSetDisplayDataType now
uses heap passed to eliminate interspersing this allocation with allocation from system heap resulting
in memory fragmentation.

Added defensive code in ODBC layer which is being flagged as most memory allocator method by gprof.