[TRAFODION-2637] Library management with Cloudera parcels
Also in this commit:

[TRAFODION-2581] Combine JVM startup code of SQL executor and UDR server

We have library management SPJs that allow a user to deploy library
files from a client machine. These SPJs make use of the pdcp command
that is installed on many clusters, but not always. The new method
uses temporary HDFS files and a TMUDF to distribute the library
file to all nodes of the cluster.

For now I left the old PUT method in place and added a new PUTFILE
method. We can remove the old method later.

A few other things fixed in the process:

- Make sure that when a UDR server calls the executor and uses a JVM,
  we share the same JVM. To do this, we need to use the same startup
  method, and I needed to change the executor startup method to accept
  the extra options used for the UDR server.

- Got rid of the "HBase compatibility mode" in JavaObjectInterface.

- A common method to execute SQL in the library management code.

- removed the obsolete init_libmgmt.sh file.

- We have special code that makes sure a UDR can request to run in
  parallel, including a request to run on all nodes. Expanded this
  special code from predefined UDFs to any UDF.