Fix for udr/TEST001 test failure
My earlier change for TRAFODION-2637 (PR #1141) caused a
non-deterministic failure of test udr/TEST001 and probably also
TEST002.

I added code that enables more parallel execution for TMUDFs.  The
tests have SELECT statements for these UDFs in them. Normally, the
logsort program sorts the output of these queries, so random
differences caused by parallel queries are ignored. However, some UDF
query have the keywords ORDER BY in them, not for the main query, but
for the inner query that is the child of the UDF. That suppressed
sorting and caused a difference.

The fix is to add an ORDER BY for the main query when we could expect
a non-deterministic order and when the child query of the UDF uses
ORDER BY.

Dave suggested that we could also change the logsort program to
recognize different types of ORDER BY, but this is probably not needed
right now.