[TRAFODION-2663] Use hbase bulkload staging dir default value
This change-set exposed a race condition in installation.

Hbase regionserver was dying because pyinstaller is starting
regionserver before the new staging dir exists in hdfs. Prior to
simplifying the settings, installer won the race of creating hdfs
dir before regionserver looked for it.

To address this, we should just use HBase default value for.

Also remove hbase setting for hbase.client.keyvalue.maxsize from
installer. That can be added to trafodion-site.xml if needed.